Enable -Wall -Wextra with -Werror and fix resulting warnings
Treat common compiler warnings as errors for project sources while
keeping vendored deps (libsodium, CLI11, httplib, nlohmann/json) exempt.
Fix unused parameters/fields, member init order, and signed/unsigned
comparisons so Clang and GCC builds stay clean.

+++ src/chain/NewUserTxHandler.cpp
@@ -198,7 +198,7 @@ chain_tx::Roe<void> NewUserTxHandler::applyNewUser(
         chain_err::E_TX_VALIDATION,
         "User account must have balance in ID_GENESIS token");
   }
-  if (it->second != tx.amount) {
+  if (it->second != static_cast<int64_t>(tx.amount)) {
     return chain_tx::TxError(
         chain_err::E_TX_VALIDATION,
         "User account must have balance in ID_GENESIS token: " +
